// client/src/common/types/pets.enums.ts
export enum EPetSpecies {
  DOG = "dog",
  CAT = "cat",
}

export enum EPetAge {
  PUPPY = "puppy",
  KITTEN = "kitten",
  ADULT = "adult",
  SENIOR = "senior",
}

export enum EPetGender {
  MALE = "male",
  FEMALE = "female",
  UNKNOWN = "unknown",
}

export enum EPetSize {
  SMALL = "small",
  MEDIUM = "medium",
  LARGE = "large",
}

export enum EPetStatus {
  AVAILABLE = "available",
  ADOPTED = "adopted",
}

export enum EAdoptionRequestStatus {
  PENDING = "pending",
  ACCEPTED = "accepted",
  CANCELED = "canceled",
}